split up roschannels into smaller .hpp files
